requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
langgraph>=0.2,<1.0
langchain>=0.3,<1.0
langchain-openai>=0.2,<1.0
pydantic-settings>=2.0,<3.0
python-dotenv>=1.0,<2.0
httpx>=0.27,<1.0
mcp>=1.0,<2.0
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
